+++ src/main/java/br/com/autorevise/mecanicagestor/api/entities/Veiculo.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.entities;
 
-import br.com.msoficinas.api.utils.conversor.UpperCaseConverter;
+import br.com.autorevise.mecanicagestor.api.utils.conversor.UpperCaseConverter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
