+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/ClienteRepository.java
@@ -1,7 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Cliente;
-import jakarta.validation.constraints.NotBlank;
+import br.com.autorevise.mecanicagestor.api.entities.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
