+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/EspecialidadeRepository.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Especialidade;
+import br.com.autorevise.mecanicagestor.api.entities.Especialidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EspecialidadeRepository extends JpaRepository<Especialidade, String> {}
