+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/FornecedorRepository.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Fornecedor;
+import br.com.autorevise.mecanicagestor.api.entities.Fornecedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
