+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/MarcaRepository.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Marca;
+import br.com.autorevise.mecanicagestor.api.entities.Marca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MarcaRepository extends JpaRepository<Marca, String> {
