+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/OrdemRepository.java
@@ -1,7 +1,7 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Estabelecimento;
-import br.com.msoficinas.api.entidades.Ordem;
+import br.com.autorevise.mecanicagestor.api.entities.Estabelecimento;
+import br.com.autorevise.mecanicagestor.api.entities.Ordem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
