+++ src/main/java/br/com/autorevise/mecanicagestor/api/repositories/ProdutoRepository.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.repositories;
 
-import br.com.msoficinas.api.entidades.Produto;
+import br.com.autorevise.mecanicagestor.api.entities.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
