+++ src/main/java/br/com/autorevise/mecanicagestor/api/web/controllers/EstoqueController.java
@@ -1,6 +1,6 @@
 package br.com.autorevise.mecanicagestor.api.web.controllers;
 
-import br.com.msoficinas.api.services.EstoqueService;
+import br.com.autorevise.mecanicagestor.api.services.EstoqueService;
 import br.com.msoficinas.api.web.request.EstoqueRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
